testing-simulator/requirements.txt
pyyaml==6.0.1
requests==2.31.0
paho-mqtt==1.6.1
chirpstack-api==4.10.2
